Rename Updatable to Actor.

+++ source/nl/colorize/multimedialib/math/Angle.java
@@ -13,21 +13,21 @@ import java.util.stream.IntStream;
  * Immutable angle in degrees, in the range between 0° and 360°. Angles are
  * normalized upon creation, so {@code new Angle(20).equals(new Angle(380)}.
  */
-public record Angle(float degrees) {
+public record Angle(double degrees) {
 
     public static final Angle ORIGIN = new Angle(0f);
     public static final List<Angle> CARDINAL = toAngles(0, 90, 180, 270);
     public static final List<Angle> INTERCARDINAL = toAngles(0, 45, 90, 135, 180, 225, 270, 315);
 
-    public Angle(float degrees) {
+    public Angle(double degrees) {
         while (degrees < 0f) {
             degrees += 360f;
         }
         this.degrees = degrees % 360f;
     }
 
-    public float getRadians() {
-        return (float) Math.toRadians(degrees);
+    public double getRadians() {
+        return Math.toRadians(degrees);
     }
 
     /**
@@ -36,8 +36,8 @@ public record Angle(float degrees) {
      * if this angle is 350 degrees, and {@code other} is 10 degrees,
      * this will return 20.
      */
-    public float distanceTo(Angle other) {
-        float phi = Math.abs(other.degrees - degrees) % 360f;
+    public double distanceTo(Angle other) {
+        double phi = Math.abs(other.degrees - degrees) % 360f;
         if (phi > 180f) {
             phi = 360f - phi;
         }
@@ -50,8 +50,8 @@ public record Angle(float degrees) {
      * {@link #distanceTo(Angle)}, but also considers the "direction" of the
      * difference and can therefore return negative values.
      */
-    public float angleTo(Angle other) {
-        float phi = other.degrees - degrees;
+    public double angleTo(Angle other) {
+        double phi = other.degrees - degrees;
 
         if (phi > 180f) {
             return phi - 360f;
@@ -66,7 +66,7 @@ public record Angle(float degrees) {
         return new Angle(degrees + other.degrees);
     }
 
-    public Angle move(float byDegrees) {
+    public Angle move(double byDegrees) {
         return new Angle(degrees + byDegrees);
     }
 
@@ -88,8 +88,8 @@ public record Angle(float degrees) {
      * specified in radians, as opposed to the constructor which requires
      * an angle specified in degrees.
      */
-    public static Angle fromRadians(float radians) {
-        return new Angle((float) Math.toDegrees(radians));
+    public static Angle fromRadians(double radians) {
+        return new Angle(Math.toDegrees(radians));
     }
 
     private static List<Angle> toAngles(int... values) {
