+++ source/nl/colorize/multimedialib/math/Size.java
@@ -9,8 +9,9 @@ package nl.colorize.multimedialib.math;
 import com.google.common.base.Preconditions;
 
 /**
- * Simple data structure for representing width and height in integer
- * precision. Negative or zero-size values are <em>not</em> permitted.
+ * Immutable representation of a size where the width and height value defined
+ * with integer precision. Negative or zero-size values are <em>not</em>
+ * permitted.
  */
 public record Size(int width, int height) {
 
@@ -19,9 +20,9 @@ public record Size(int width, int height) {
         Preconditions.checkArgument(height > 0, "Invalid height: " + height);
     }
 
-    public Size multiply(float factor) {
+    public Size multiply(double factor) {
         Preconditions.checkArgument(factor > 0f, "Invalid factor: " + factor);
-        return new Size(Math.round(width * factor), Math.round(height * factor));
+        return new Size((int) Math.round(width * factor), (int) Math.round(height * factor));
     }
 
     @Override
