+++ source/nl/colorize/multimedialib/math/Sphere.java
@@ -9,10 +9,10 @@ package nl.colorize.multimedialib.math;
 import com.google.common.base.Preconditions;
 
 /**
- * Immutable three-dimensional sphere with float precision. The sphere is
- * defined based on the location of its center point and its radius.
+ * Immutable three-dimensional sphere defined by its center point and its
+ * radius around the center point.
  */
-public record Sphere(Point3D center, float radius) implements Shape3D {
+public record Sphere(Point3D center, double radius) implements Shape3D {
 
     public Sphere {
         Preconditions.checkArgument(radius >= 0f, "Invalid radius: " + radius);
@@ -21,7 +21,7 @@ public record Sphere(Point3D center, float radius) implements Shape3D {
     /**
      * Creates a sphere with its center point located at the origin (0, 0, 0).
      */
-    public Sphere(float radius) {
+    public Sphere(double radius) {
         this(Point3D.ORIGIN, radius);
     }
